+++ c/src/hashtree_private.h
@@ -83,10 +83,10 @@ struct _tree_tree
     /* Type specific functions used by table. */
     uint16_t(*hash_key)(void *);
     uint16_t(*hash_key_list)(void *);
-    int16_t(*key_compare)(void *,void *);
+    int32_t(*key_compare)(void *,void *);
     void*(*key_copy)(void *);
     void(*key_free)(void *);
-    int16_t(*data_compare)(void *,void *);
+    int32_t(*data_compare)(void *,void *);
     void*(*data_copy)(void *);
     void(*data_free)(void *);
